+++ CHANGELOG.md
@@ -59,6 +59,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added vibration feedback support for XInput gamepads and DualShock 4.
 - Added the light bar color feedback support for DualShock 4. The color changes depending from health and oxygen.
 - Passport text box height is adjusted depending from HUD scale.
+- Some legacy options are removed (Triple Buffer, Perspective Correction, Texel Adjustment, Disable 16 bit textures, Don't sort transparent polys).
 
 ### The original game bugfixes
 - Fixed a bug that prevented the display of the save counter until the game relaunch, if the game was saved in an empty slot.